@@ -25,7 +25,7 @@
 	"require": {
 		"maxmind/minfraud": "^1.4",
 		"geoip2/geoip2": "^2.9",
-		"monolog/monolog": "~1.18",
+		"monolog/monolog": "~1.22.1",
 		"neitanod/forceutf8": "^2.0",
 		"php": ">=5.4",
 		"predis/predis": "^1.1",
@@ -39,7 +39,7 @@
 	"require-dev": {
 		"jakub-onderka/php-parallel-lint": "1.0.0",
 		"jakub-onderka/php-console-highlighter": "0.3.2",
-		"mediawiki/mediawiki-codesniffer": "19.1.0",
+		"mediawiki/mediawiki-codesniffer": "25.0.0",
 		"wikimedia/testing-access-wrapper": "~1.0",
 		"mediawiki/minus-x": "0.3.1"
 	},
@@ -53,7 +53,7 @@
 		"optimize-autoloader": true,
 		"prepend-autoloader": false,
 		"platform": {
-			"php": "5.6.36"
+			"php": "7.0"
 		}
 	},
 	"scripts": {
