@@ -154,8 +154,31 @@ class AdyenAdapter extends GatewayAdapter {
 	}
 
 	function doPayment() {
+		$apiResult = $this->do_transaction( 'donate' );
+		$data = $apiResult->getData();
+		// FIXME: Ugh, weird key inherited from legacy GlobalCollect
+		if ( !empty( $data['FORMACTION'] ) && !empty( $data['gateway_params'] ) ) {
+			// We've got a URL to send them to and data to post.
+			// Now figure out if we're using an iframe or a redirect
+			// For the Adyen workflow, we do a full redirect for all iOS
+			// and Safari browsers. We decide this on the front end and
+			// post back a 'processor_form' value that will match either
+			// the 'base' or 'redirect' skin code.
+			$processorForm = $this->getData_Unstaged_Escaped( 'processor_form' );
+			$skinCodes = $this->getSkinCodes();
+			if ( $processorForm === $skinCodes['redirect'] ) {
+				return PaymentResult::newRedirect(
+					$data['FORMACTION'], $data['gateway_params']
+				);
+			}
+			return PaymentResult::newIframe(
+				$data['FORMACTION'], $data['gateway_params']
+			);
+		}
+		// If we've fallen through to here, there must be some problem. Use the
+		// default instantiation function.
 		return PaymentResult::fromResults(
-			$this->do_transaction( 'donate' ),
+			$apiResult,
 			$this->getFinalStatus()
 		);
 	}
@@ -206,6 +229,7 @@ class AdyenAdapter extends GatewayAdapter {
 						'FORMACTION' => $formaction,
 						'gateway_params' => $requestParams,
 					] );
+					// FIXME: might be an iframe, might be a redirect
 					$this->logger->info(
 						"launching external iframe request: " . print_r( $requestParams, true )
 					);
@@ -262,10 +286,10 @@ class AdyenAdapter extends GatewayAdapter {
 		$this->addRequestData( [
 			'order_id' => $orderId,
 		] );
-		$gateway_txn_id = isset( $requestValues['pspReference'] ) ? $requestValues['pspReference'] : '';
+		$gateway_txn_id = $requestValues['pspReference'] ?? '';
 		$this->transaction_response->setGatewayTransactionId( $gateway_txn_id );
 
-		$result_code = isset( $requestValues['authResult'] ) ? $requestValues['authResult'] : '';
+		$result_code = $requestValues['authResult'] ?? '';
 		$paymentResult = null;
 		if ( $result_code == 'PENDING' || $result_code == 'AUTHORISED' ) {
 			// Both of these are listed as pending because we have to submit a capture
