@@ -6,23 +6,31 @@
 		safari = window.safari !== undefined;
 
 	function redirect( result ) {
-		var $pForm, $payment = $( '#payment-form' );
+		var $pForm, $payment = $( '#payment-form' ),
+			// TODO: delete formaction and gateway_params
+			// after caches are all clear
+			url = result.redirect || result.formaction,
+			params = result.formData || result.gateway_params;
 		$pForm = $(
 			'<form></form>', {
 				method: 'post',
-				action: result.formaction,
+				action: url,
 				id: 'submit-payment'
 			}
 		);
-		populateHiddenFields( result.gateway_params, $pForm );
+		populateHiddenFields( params, $pForm );
 		$payment.append( $pForm );
 
-		$pForm.prop( 'action', result.formaction );
+		$pForm.prop( 'action', url );
 		$pForm.submit();
 	}
 
 	function showIframe( result ) {
-		var $pForm, $payment = $( '#payment-form' );
+		var $pForm, $payment = $( '#payment-form' ),
+			// TODO: delete formaction and gateway_params
+			// after deploying 'donation' API update
+			url = result.iframe || result.formaction,
+			params = result.formData || result.gateway_params;
 
 		// Empty the div; add the target iframe; then submit the request for the iframe contents
 		$payment.append(
@@ -40,12 +48,12 @@
 		$pForm = $(
 			'<form></form>', {
 				method: 'post',
-				action: result.formaction,
+				action: url,
 				target: 'adyen-iframe',
 				id: 'fetch-iframe-form'
 			}
 		);
-		populateHiddenFields( result.gateway_params, $pForm );
+		populateHiddenFields( params, $pForm );
 		$payment.append( $pForm );
 
 		$payment.find( '#fetch-iframe-form' ).submit();
