@@ -14,9 +14,8 @@ class AstroPayStatusQuery extends Maintenance {
 	public function __construct() {
 		parent::__construct();
 
-		if ( method_exists( $this, 'requireExtension' ) ) {
-			$this->requireExtension( 'Donation Interface' );
-		}
+		$this->requireExtension( 'Donation Interface' );
+
 		$this->addArg( 'id', 'Contribution tracking ID', true );
 	}
 
