@@ -10,7 +10,8 @@
 
 		var $form = $( '<iframe>' )
 			.attr( {
-				src: result.formaction,
+				// TODO: remove formaction after deploying donation api change
+				src: result.iframe || result.formaction,
 				width: 318,
 				height: 316,
 				frameborder: 0,
@@ -20,11 +21,15 @@
 		$( '#payment-form' ).append( $form );
 	}
 
-	if ( di.forms.isIframe() ) {
-		di.forms.submit = function () {
-			di.forms.callDonateApi( function ( result ) {
-				showIframe( result );
-			} );
-		};
+	function handleResult( result ) {
+		if ( di.forms.isIframe() && !result.redirect ) {
+			showIframe( result );
+		}
+		location.replace( result.redirect );
 	}
+
+	di.forms.submit = function () {
+		di.forms.callDonateApi( handleResult );
+	};
+
 } )( jQuery, mediaWiki );
