@@ -16,9 +16,8 @@ class IngenicoGetOrderStatusMaintenance extends Maintenance {
 	public function __construct() {
 		parent::__construct();
 
-		if ( method_exists( $this, 'requireExtension' ) ) {
-			$this->requireExtension( 'Donation Interface' );
-		}
+		$this->requireExtension( 'Donation Interface' );
+
 		$this->addOption( 'file', 'Read order IDs in from a file',
 			true, true, 'f' );
 	}
