@@ -2,7 +2,7 @@
 
 class PayPalCountry implements UnstagingHelper {
 
-	static $nonStandardCodes = [
+	private static $nonStandardCodes = [
 		'C2' => 'CN', // mutant China code for merchants outside of Chine
 		'AN' => 'NL', // Netherlands Antilles is part of Netherlands since 2010
 	];
